Move all multiproc to unified and add to CI
