+++ .github/workflows/run-tests.yml
@@ -105,3 +105,7 @@ jobs:
       - name: Run MISC tests
         run: |
           ${{ env.TEST_OUTPUT_DIR }}/umd/misc/umd_misc_tests
+
+      - name: Run unified tests
+        run: |
+          ${{ env.TEST_OUTPUT_DIR }}/umd/unified/unified_tests
