+++ tests/api/CMakeLists.txt
@@ -9,7 +9,6 @@ set(API_TESTS_SRCS
     test_tlb_manager.cpp
     test_tt_device.cpp
     test_software_harvesting.cpp
-    test_multiprocess.cpp
 )
 
 add_executable(api_tests ${API_TESTS_SRCS})
